+++ CardGame/Model/QuestionData.swift
@@ -1,0 +1,15 @@
+//
+//  QuestionData.swift
+//  CardGame
+//
+//  Created by Grzegorz Mzyk on 04/07/2024.
+//
+
+import Foundation
+class QuestionDataModel {
+    
+    static let questions: [Questions] = [
+        Questions(question: "pytanie 1", punishment: 2, points: 1),
+        Questions(question: "pytanie 2", punishment: 3, points: 1)
+    ]
+}
